python/05-Web开发/examples/fastapi_min/pyproject.toml
[project]
name = "fastapi-min"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.19",
  "pydantic-settings>=2.3",
  "APScheduler>=3.10",
  "PyJWT>=2.8",
  "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "requests",
  "httpx",
  "pytest-asyncio",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


